Remove __is_vectorizable_range due to compiler optimization
This utility was added to ensure that random access but non contiguous
ranges (e.g. reverse_iterator) do not go through vectorization paths.
However, after further experimentation, the oneAPI compiler is capable
of accounting for this and enabling vectorization.

For other non-contiguous types that go through global memory (e.g.
permutation iterator), the access pattern is already poor, so
performance degradation due to the change in access pattern is a non-issue here.

Signed-off-by: Matthew Michel <matthew.michel@intel.com>